+++ src/Module/Migration/ModuleFile10To20Migration.php
@@ -42,6 +42,25 @@ class ModuleFile10To20Migration implements JsonMigration
             $data->modules = $data->packages;
             unset($data->packages);
         }
+
+        if (isset($data->{'path-mappings'})) {
+            $data->resources = $data->{'path-mappings'};
+            unset($data->{'path-mappings'});
+        }
+
+        if (isset($data->override) && !is_array($data->override)) {
+            $data->override = array($data->override);
+        }
+
+        if (isset($data->{'override-order'})) {
+            $data->order = $data->{'override-order'};
+            unset($data->{'override-order'});
+        }
+
+        if (isset($data->override)) {
+            $data->depend = $data->override;
+            unset($data->override);
+        }
     }
 
     public function down(stdClass $data)
@@ -53,6 +72,25 @@ class ModuleFile10To20Migration implements JsonMigration
         if (isset($data->modules)) {
             $data->packages = $data->modules;
             unset($data->modules);
+        }
+
+        if (isset($data->resources)) {
+            $data->{'path-mappings'} = $data->resources;
+            unset($data->resources);
+        }
+
+        if (isset($data->override) && 1 === count($data->override)) {
+            $data->override = reset($data->override);
+        }
+
+        if (isset($data->order)) {
+            $data->{'override-order'} = $data->order;
+            unset($data->order);
+        }
+
+        if (isset($data->depend)) {
+            $data->override = $data->depend;
+            unset($data->depend);
         }
     }
 }