+++ tests/Module/ModuleFileConverterTest.php
@@ -85,7 +85,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $moduleFile->addTypeDescriptor(new BindingTypeDescriptor($type, 'Description of my type.', array(
             'param' => 'Description of the parameter.',
         )));
-        $moduleFile->setOverriddenModules(array('acme/blog'));
+        $moduleFile->setDependencies(array('acme/blog'));
         $moduleFile->setExtraKeys(array(
             'extra1' => 'value',
             'extra2' => (object) array('key' => 'value'),
@@ -94,7 +94,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $jsonData = (object) array(
             '$schema' => 'http://puli.io/schema/2.0/manager/module',
             'name' => 'my/application',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => 'res',
             ),
             'bindings' => (object) array(
@@ -120,7 +120,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
                     ),
                 ),
             ),
-            'override' => 'acme/blog',
+            'depend' => array('acme/blog'),
             'extra' => (object) array(
                 'extra1' => 'value',
                 'extra2' => (object) array(
@@ -177,7 +177,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
 
         $jsonData = (object) array(
             '$schema' => 'http://puli.io/schema/2.0/manager/module',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/vendor/a' => 'foo',
                 '/vendor/b' => 'foo',
                 '/vendor/c' => 'foo',
@@ -428,7 +428,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
 
         $jsonData = (object) array(
             '$schema' => 'http://puli.io/schema/2.0/manager/module',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => array('res', 'assets'),
             ),
         );
@@ -439,11 +439,11 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
     public function testToJsonMultipleOverriddenModules()
     {
         $moduleFile = new ModuleFile();
-        $moduleFile->setOverriddenModules(array('acme/blog1', 'acme/blog2'));
+        $moduleFile->setDependencies(array('acme/blog1', 'acme/blog2'));
 
         $jsonData = (object) array(
             '$schema' => 'http://puli.io/schema/2.0/manager/module',
-            'override' => array(
+            'depend' => array(
                 'acme/blog1',
                 'acme/blog2',
             ),
@@ -456,7 +456,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
     {
         $jsonData = (object) array(
             'name' => 'my/application',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => 'res',
             ),
             'bindings' => (object) array(
@@ -482,7 +482,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
                     ),
                 ),
             ),
-            'override' => 'acme/blog',
+            'depend' => array('acme/blog'),
             'extra' => (object) array(
                 'extra1' => 'value',
                 'extra2' => (object) array(
@@ -517,7 +517,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(array(new BindingTypeDescriptor($type, 'Description of my type.', array(
             'param' => 'Description of the parameter.',
         ))), $moduleFile->getTypeDescriptors());
-        $this->assertSame(array('acme/blog'), $moduleFile->getOverriddenModules());
+        $this->assertSame(array('acme/blog'), $moduleFile->getDependencies());
         $this->assertEquals(array(
             'extra1' => 'value',
             'extra2' => (object) array('key' => 'value'),
@@ -553,7 +553,7 @@ class ModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $this->assertNull($moduleFile->getModuleName());
         $this->assertSame(array(), $moduleFile->getPathMappings());
         $this->assertSame(array(), $moduleFile->getBindingDescriptors());
-        $this->assertSame(array(), $moduleFile->getOverriddenModules());
+        $this->assertSame(array(), $moduleFile->getDependencies());
     }
 
     public function testFromJsonBindingTypeWithRequiredParameter()