+++ tests/Module/RootModuleFileConverterTest.php
@@ -26,7 +26,6 @@ use Puli\Manager\Api\Module\InstallInfo;
 use Puli\Manager\Api\Module\RootModuleFile;
 use Puli\Manager\Api\PuliPlugin;
 use Puli\Manager\Api\Repository\PathMapping;
-use Puli\Manager\Module\ModuleFileConverter;
 use Puli\Manager\Module\RootModuleFileConverter;
 use Puli\Manager\Tests\Discovery\Fixtures\Bar;
 use Puli\Manager\Tests\Discovery\Fixtures\Foo;
@@ -57,7 +56,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
     private $versioner;
 
     /**
-     * @var ModuleFileConverter
+     * @var RootModuleFileConverter
      */
     private $converter;
 
@@ -92,8 +91,8 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $moduleFile->addTypeDescriptor(new BindingTypeDescriptor($type, 'Description of my type.', array(
             'param' => 'Description of the parameter.',
         )));
-        $moduleFile->setOverriddenModules(array('acme/blog'));
-        $moduleFile->setOverrideOrder(array(
+        $moduleFile->setDependencies(array('acme/blog'));
+        $moduleFile->setModuleOrder(array(
             'acme/blog-extension1',
             'acme/blog-extension2',
         ));
@@ -116,7 +115,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $jsonData = (object) array(
             '$schema' => 'http://puli.io/schema/2.0/manager/module',
             'name' => 'my/application',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => 'res',
             ),
             'bindings' => (object) array(
@@ -142,14 +141,14 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
                     ),
                 ),
             ),
-            'override' => 'acme/blog',
+            'depend' => array('acme/blog'),
             'extra' => (object) array(
                 'extra1' => 'value',
                 'extra2' => (object) array(
                     'key' => 'value',
                 ),
             ),
-            'override-order' => array(
+            'order' => array(
                 'acme/blog-extension1',
                 'acme/blog-extension2',
             ),
@@ -292,7 +291,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
     {
         $jsonData = (object) array(
             'name' => 'my/application',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => 'res',
             ),
             'bindings' => (object) array(
@@ -318,14 +317,14 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
                     ),
                 ),
             ),
-            'override' => 'acme/blog',
+            'depend' => array('acme/blog'),
             'extra' => (object) array(
                 'extra1' => 'value',
                 'extra2' => (object) array(
                     'key' => 'value',
                 ),
             ),
-            'override-order' => array(
+            'order' => array(
                 'acme/blog-extension1',
                 'acme/blog-extension2',
             ),
@@ -400,12 +399,12 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(array(new BindingTypeDescriptor($type, 'Description of my type.', array(
             'param' => 'Description of the parameter.',
         ))), $moduleFile->getTypeDescriptors());
-        $this->assertSame(array('acme/blog'), $moduleFile->getOverriddenModules());
+        $this->assertSame(array('acme/blog'), $moduleFile->getDependencies());
         $this->assertEquals(array(
             'extra1' => 'value',
             'extra2' => (object) array('key' => 'value'),
         ), $moduleFile->getExtraKeys());
-        $this->assertSame(array('acme/blog-extension1', 'acme/blog-extension2'), $moduleFile->getOverrideOrder());
+        $this->assertSame(array('acme/blog-extension1', 'acme/blog-extension2'), $moduleFile->getModuleOrder());
         $this->assertEquals(array($installInfo1, $installInfo2), $moduleFile->getInstallInfos());
         $this->assertSame('puli-dir', $config->get(Config::PULI_DIR));
         $this->assertSame('Puli\MyFactory', $config->get(Config::FACTORY_OUT_CLASS));
@@ -445,8 +444,8 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $this->assertNull($moduleFile->getModuleName());
         $this->assertSame(array(), $moduleFile->getPathMappings());
         $this->assertSame(array(), $moduleFile->getBindingDescriptors());
-        $this->assertSame(array(), $moduleFile->getOverriddenModules());
-        $this->assertSame(array(), $moduleFile->getOverrideOrder());
+        $this->assertSame(array(), $moduleFile->getDependencies());
+        $this->assertSame(array(), $moduleFile->getModuleOrder());
     }
 
     public function testFromJsonInheritsBaseConfig()